+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * Render a pkgdown website (PR #13).
 
+* Add `to_string()` and `print()` methods to the `Record` class and (incomplete) `describe()` method to the `Artifact()` class (PR #22)
+
 ## MAJOR CHANGES
 
 * Refactored the internal class data structures for better modularity and extensibility (PR #8).